+++ styles/Home.module.css
@@ -1,49 +1,129 @@
-.container {
-  padding: 0 2rem;
+/* fonts */
+@import url('https://fonts.googleapis.com/css2?family=Raleway:wght@400;600;800;900&display=swap');
+
+/* general styles */
+
+.page_width {
+  max-width: 1400px;
+  /* width: 100%; */
+  margin: 0 auto;
+}
+
+.h2 {
+  font-family: 'Raleway', sans-serif;
+  color: var(--primary);
+  font-size: 2.5rem;
+  margin-bottom: 0;
+}
+.logo_container {
+  text-align: center;
+  margin-bottom: 1rem;
 }
 
 .main {
-  max-width: 90%;
-  margin-left: 5%;
-  margin-right: 5%;
-  margin-top: 0%;
-  min-height: 60vh;
-  flex: 1;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-evenly;
-  align-items: center;
+  margin: 1rem;
 }
 
-.map_container{
-    border-radius: 10px;
-    width:40vw;
-    height: 60vh;
-
+img .logo_image {
+  height: 300px;
+  width: 200px;
+}
+.main .about_us_main_grid_container {
+  width: 100%;
+  margin: 0 auto;
+  display: grid;
+  grid-template-columns: repeat(12, 1fr);
+  gap: 1.8rem;
 }
 
-
-
-.filter_bar{
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
-  background-color: #225544;
- 
+.about_us_description_container {
+  background-color: var(--secondary);
+  /* display: flex;
+  flex-direction: column;
+  width: 50%;
+  height: 50%;
+  margin: 3rem;
+  margin-right: 1rem;
+  border-radius: 10px; */
+  color: white;
+  font-family: 'Raleway', sans-serif;
+  grid-column: span 6;
+  padding: 0.5rem 2rem 0rem 2rem;
   border-radius: 10px;
-  max-width: 87%;
-  margin: 1vh auto;
-  margin-top:2vh ;
-
 }
 
+.about_us_description_paragraph {
+  /* margin-left: 1.5rem;
+  margin-right: 1.5rem; */
+  font-size: 1.1rem;
+  font-family: 'Raleway', serif;
+}
 
+.about_us_images_container {
+  background-color: var(--secondary);
+  /* display: flex;
+  flex-direction: row;
+  width: 50%;
+  height: 50%;
+  margin: 3rem; */
+  /* margin-left: 1rem; */
+  /* border-radius: 10px; */
+  /* flex-wrap: wrap;
+  justify-content: space-around; */
+  padding: 1rem;
+  grid-column: span 6;
+  align-self: center;
+  border-radius: 10px;
+}
 
+.images_grid_container {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  align-self: center;
+  padding: 1rem;
 
+  /* grid-row: 400px; */
+  gap: 1rem;
 
+  /* margin: auto; */
+  /* gap: 1.5rem; */
+}
+.images {
+  width: 100%;
+  height: 100%;
+  box-shadow: rgba(50, 50, 93, 0.25) 0px 6px 12px -2px,
+    rgba(0, 0, 0, 0.3) 0px 3px 7px -3px;
+  border-radius: 4px;
+  object-fit: cover;
+  margin: 5px;
+  align-self: center;
+}
 
+.review_section {
+  display: grid;
+  grid-template-columns: repeat(10, 1fr);
+  gap: 1rem;
+}
 
+.reviews {
+  margin-top: 2rem;
+  background-color: var(--secondary);
+  /* display: flex;
+  flex-direction: column;
+  width: 50%;
+  margin: 2rem;
+  margin-right: 1rem;
+  border-radius: 10px; */
+  color: white;
+  font-family: 'Raleway', sans-serif;
+  padding: 1rem;
+  grid-column: span 2;
+  gap: 1rem;
+  border-radius: 10px;
+}
 
-
-
-
+.reviews h2 {
+  margin-top: 0;
+  color: var(--primary);
+  font-size: 1.8rem;
+}
