+++ styles/Locations.module.css
@@ -1,0 +1,49 @@
+.container {
+  padding: 0 2rem;
+}
+
+.main {
+  max-width: 90%;
+  margin-left: 5%;
+  margin-right: 5%;
+  margin-top: 0%;
+  min-height: 60vh;
+  flex: 1;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  align-items: center;
+}
+
+.map_container{
+    border-radius: 10px;
+    width:40vw;
+    height: 60vh;
+
+}
+
+
+
+.filter_bar{
+  display: flex;
+  justify-content: space-evenly;
+  align-items: center;
+  background-color: #225544;
+ 
+  border-radius: 10px;
+  max-width: 87%;
+  margin: 1vh auto;
+  margin-top:2vh ;
+
+}
+
+
+
+
+
+
+
+
+
+
+
